requirements.txt
python-telegram-bot>=20,<21
python-dotenv
 
